refactor(ui): rename sonner.tsx to toaster.tsx
The file fully migrated from the sonner library to react-toastify in a
previous PR, but the filename was left as sonner.tsx. Rename it to
toaster.tsx to match the exported `Toaster` component and the
react-toastify backing, and update the single import in App.tsx.
