+++ src/App.tsx
@@ -11,7 +11,7 @@ import IndexPage from '@/pages'
 import ErrorPage from '@/pages/error'
 import InitPage from '@/pages/init'
 import { PersistentPageLayout } from '@/shared/layout/PersistentPageLayout'
-import { Toaster } from '@/shared/ui/sonner'
+import { Toaster } from '@/shared/ui/toaster'
 import '@/styles/global.css'
 import { useEffect } from 'react'
 import { Route, Routes } from 'react-router'
